changed float to int

+++ pt2phoneProfitCalculator.py
@@ -9,7 +9,7 @@ def dailyProfit():
     totalProfit = 0
 
     #get product category input
-    categoryInput = float(input("Enter product number 1-5, or enter 0 to stop: \n"))
+    categoryInput = int(input("Enter product number 1-5, or enter 0 to stop: \n"))
 
     while categoryInput != 0: 
     #error message for invalid input
@@ -17,27 +17,27 @@ def dailyProfit():
             print("Invalid input, please enter a valid number.")
         else:
             if categoryInput ==  1:
-                quantityInput = float(input("Enter quantity sold: \n"))
+                quantityInput = int(input("Enter quantity sold: \n"))
                 profit = productCategory[1] * quantityInput #access the value in the dictionary by using the key
                 totalProfit += profit #add the profit to the total amount of profit
             elif categoryInput == 2:
-                quantityInput = float(input("Enter quantity sold: \n"))
+                quantityInput = int(input("Enter quantity sold: \n"))
                 profit = productCategory[2] * quantityInput
                 totalProfit += profit
             elif categoryInput == 3:
-                quantityInput = float(input("Enter quantity sold: \n"))
+                quantityInput = int(input("Enter quantity sold: \n"))
                 profit = productCategory[3] * quantityInput
                 totalProfit += profit
             elif categoryInput == 4:
-                quantityInput = float(input("Enter quantity sold: \n"))
+                quantityInput = int(input("Enter quantity sold: \n"))
                 profit = productCategory[4] * quantityInput
                 totalProfit += profit
             elif categoryInput == 5:
-                quantityInput = float(input("Enter quantity sold: \n"))
+                quantityInput = int(input("Enter quantity sold: \n"))
                 profit = productCategory[5] * quantityInput
                 totalProfit += profit
         #continue to ask for product category
-        categoryInput = float(input("Enter product number 1-5, or enter 0 to stop: \n"))
+        categoryInput = int(input("Enter product number 1-5, or enter 0 to stop: \n"))
     return totalProfit
